Change how markers are added to map
New wrapper class was added: QMMarker
It contains the position of the marker, type, color, id and name of it

Could not find a way to call javascript functions passing a QObject directly
as argument. So when we want to add a marker to the map, we first add it to QMMapView
and pass its id to javascript. In javascript, a Q_INVOKABLE function can be called
to get the C++ marker class and get its attributes

Signed-off-by: Thiago de Barros Lacerda <thiagotbl@gmail.com>